requirements.txt
numpy
pillow
tensorflow
streamlit
streamlit-option-menu
opencv-python-headless